d_a_npc_aj1: createInit s32->u8 (retail caller clrlwi.24 in _create) — _create 99.26->99.66 (remaining = rodata/a_siz_tbl anchors); fresh .o postdates

+++ include/d/actor/d_a_npc_aj1.h
@@ -28,7 +28,7 @@ public:
     u8 init_AJ1_0();
     u8 init_AJ1_1();
     u8 init_AJ1_2();
-    s32 createInit();
+    u8 createInit();
     void play_animation();
     void ctrl_WAITanm();
     void ctrl_TIREanm();
